Cranelift: Adjust virtual SP after `tail` call-conv callees return (#6586)
* Cranelift: Adjust virtual SP after `tail` call-conv callees return

Callees that use the `tail` calling convention will pop stack arguments from the
stack for their callers. They do not, however, adjust the caller's virtual SP,
so that still needs to happen in our ABI and `CallSite` code. This is, however,
slightly trickier than just emitting a nominal SP adjustment pseudo-instruction
because we cannot let regalloc attempt to spill or reload values between the
call and the SP adjustment because the stack offsets will be off by the size of
the stack arguments to the call. Therefore, we add the number of bytes that the
callee pops to the `CallInfo` structures and have emission update the virtual SP
atomically with regards to the call itself.

Fixes #6581
Fixes #6582

Co-Authored-By: Jamey Sharp <jsharp@fastly.com>

* Cranelift: Have `fuzzgen` generate functions with the `tail` calling convention

---------

Co-authored-by: Jamey Sharp <jsharp@fastly.com>