winch: Remove stack pointer check at callsite (#6664)
This commits removes an assert that checked that the stack pointer
position at the end of a call should be greater or equal than the
position registered at the callsite.

Even though this is true in most cases, there are cases in which this is
invariant is not met and as well as there are cases in which the stack pointer will
inevitably be greater than the position registered at callsite:

1. When the call setup doesn't spill any values and instead it
   only consumes memory values from the value stack, the stack pointer
   can end up being less than what it was at the callsite.

2. When the call setup spills values that are not going to be consumed
   by the call (not used as params to the function) the stack pointer
   position can end up being greater than what it was at the callsite.

The assert was originally introduced to ensure the right deallocation of
stack space consumed by the call, and it could be improved by applying
the heuristics mentioned above, but I prefer to remove it since we
already assert when emitting the epilogue that both the value stack and
machine stack are in the correct state when fishing compilation.

This change includes an extra test in which the original invariant
doesn't hold (case 2 described above occurs).