Dependency gardening for Wasmtime (#6731)
* Remove deny.toml exception for wasm-coredump-builder

This isn't used any more so no need to continue to list this.

* Update Wasmtime's pretty_env_logger dependency

This removes a `deny.toml` exception for that crate, but `openvino-sys`
still depends on `pretty_env_logger 0.4.0` so a new exception is added
for that.

* Update criterion and clap dependencies

This commit started out by updating the `criterion` dependency to remove
an entry in `deny.toml`, but that ended up transitively requiring a
`clap` dependency upgrade from 3.x to 4.x because `criterion` uses
pieces of clap 4.x. Most of this commit is then dedicated to updating
clap 3.x to 4.x which was relatively simple, mostly renaming attributes
here and there.

* Update gimli-related dependencies

I originally wanted to remove the `indexmap` clause in `deny.toml` but
enough dependencies haven't updated from 1.9 to 2.0 that it wasn't
possible. In the meantime though this updates some various dependencies
to bring them to the latest and a few of them now use `indexmap` 2.0.

* Update deps to remove `windows-sys 0.45.0`

This involved updating tokio/mio and then providing new audits for new
crates. The tokio exemption was updated from its old version to the new
version and tokio remains un-audited.

* Update `syn` to 2.x.x

This required a bit of rewriting for the component-macro related bits
but otherwise was pretty straightforward. The `syn` 1.x.x track is still
present in the wasi-crypto tree at this time.

I've additionally added some trusted audits for my own publications of
`wasm-bindgen`

* Update bitflags to 2.x.x

This updates Wasmtime's dependency on the `bitflags` crate to the 2.x.x
track to keep it up-to-date.

* Update the cap-std family of crates

This bumps them all to the next major version to keep up with updates.
I've additionally added trusted entries for publishes of cap-std crates
from Dan.

There's still lingering references to rustix 0.37.x which will need to
get weeded out over time.

* Update memoffset dependency to latest

Avoids having two versions in our crate graph.

* Fix tests

* Update try_from for wiggle flags

* Fix build on AArch64 Linux

* Enable `event` for rustix on Windows too